+++ cli/src/commands/exec.rs
@@ -42,7 +42,7 @@ impl Opts {
 			ctx,
 			cmd::RunWithRivetOpts {
 				command: &self.command,
-				envs: Vec::new(),
+				env: Vec::new(),
 				namespace: self.namespace.as_deref(),
 				token,
 			},